// ExistingCode/src/data-access/domain/tables/drillmethod/schema.ts
export const DrillMethodTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	drillMethodId: z.any().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	collarId: z.any(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: decimal
	 */
	depthFrom: z.number(),
	/**
	 * dataType: decimal
	 */
	depthTo: z.number(),
	/**
	 * dataType: datetimeoffset
	 */
	startDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	endDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	driller1: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	driller2: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillSize: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillRigType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillCompany: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	sampleType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});